requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
boto==2.49.0
boto3==1.26.151
botocore==1.29.151
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.1
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-countries==7.5.1
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-timezone-field==5.0
django-widget-tweaks==1.4.12
idna==3.4
jmespath==1.0.1
kavenegar==1.1.2
kombu==5.3.0
passlib==1.7.4
Pillow==9.5.0
prompt-toolkit==3.0.38
pycparser==2.21
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
